requirements.txt
pandas
ipython
pyarrow
pyyaml
py-radix